+++ Holovibes/includes/batch_input_queue.hh
@@ -41,8 +41,6 @@ class BatchInputQueue : public DisplayQueue
   public: /* Public methods */
     BatchInputQueue(const uint total_nb_frames, const uint batch_size, const camera::FrameDescriptor& fd);
 
-    BatchInputQueue(const uint total_nb_frames, const camera::FrameDescriptor& fd);
-
     ~BatchInputQueue();
 
     /*! \brief Enqueue a frame in the queue