+++ Holovibes/includes/enum/enum_composite_kind.hh
@@ -4,6 +4,8 @@
  */
 #pragma once
 
+#include <map>
+
 namespace holovibes
 {
 /*! \enum CompositeKind
@@ -15,4 +17,15 @@ enum class CompositeKind
     RGB = 0, /*!< Composite in RGB */
     HSV      /*!< Composite in HSV */
 };
+
+static std::map<CompositeKind, std::string> composite_kind_to_string = {
+    {CompositeKind::RGB, "RGB"},
+    {CompositeKind::HSV, "HSV"},
+};
+
+static std::map<std::string, CompositeKind> string_to_composite_kind = {
+    {"RGB", CompositeKind::RGB},
+    {"HSV", CompositeKind::HSV},
+};
+
 } // namespace holovibes