+++ Holovibes/includes/enum/enum_computation.hh
@@ -4,6 +4,8 @@
  */
 #pragma once
 
+#include <map>
+
 namespace holovibes
 {
 /*! \enum Computation
@@ -15,4 +17,15 @@ enum class Computation
     Raw = 0, /*!< Interferogram recorded */
     Hologram /*!<  Reconstruction of the object */
 };
+
+static std::map<std::string, Computation> string_to_computation = {
+    {"RAW", Computation::Raw},
+    {"HOLOGRAM", Computation::Hologram},
+};
+
+static std::map<Computation, std::string> computation_to_string = {
+    {Computation::Raw, "RAW"},
+    {Computation::Hologram, "HOLOGRAM"},
+};
+
 } // namespace holovibes