+++ Holovibes/sources/cli.cc
@@ -4,7 +4,7 @@
 
 #include "tools.hh"
 #include "icompute.hh"
-#include "ini_config.hh"
+#include "holovibes_config.hh"
 #include "input_frame_file_factory.hh"
 #include "enum_record_mode.hh"
 #include "global_state_holder.hh"
@@ -41,9 +41,6 @@ static void print_verbose(const holovibes::OptionsDescriptor& opts, const holovi
 {
     std::cout << "Config:\n";
     boost::property_tree::ptree ptree;
-    // Ask if removable, cli should not change settings
-    // holovibes::ini::save_ini(cd, opts);
-    // boost::property_tree::write_ini(std::cout, ptree);
     std::cout << std::endl;
 
     std::cout << "Input file: " << opts.input_path.value() << "\n";
@@ -69,7 +66,8 @@ bool get_first_and_last_frame(const holovibes::OptionsDescriptor& opts,
                               const uint& nb_frames,
                               holovibes::ComputeDescriptor& cd)
 {
-    auto err_message = [&](const std::string& name, const uint& value, const std::string& option) {
+    auto err_message = [&](const std::string& name, const uint& value, const std::string& option)
+    {
         std::cerr << option << " (" << name << ") value: " << value
                   << " is not valid. The valid condition is: 1 <= " << name
                   << " <= nb_frame. For this file nb_frame = " << nb_frames << ".";
@@ -109,7 +107,8 @@ static bool set_parameters(holovibes::Holovibes& holovibes, const holovibes::Opt
     holovibes::io_files::InputFrameFile* input_frame_file =
         holovibes::io_files::InputFrameFileFactory::open(input_path);
 
-    input_frame_file->import_compute_settings(holovibes.get_cd());
+    if (!opts.compute_settings_path)
+        input_frame_file->import_compute_settings(holovibes::api::get_cd());
 
     const camera::FrameDescriptor& fd = input_frame_file->get_frame_descriptor();
 
@@ -190,11 +189,11 @@ int start_cli(holovibes::Holovibes& holovibes, const holovibes::OptionsDescripto
 {
     auto& cd = holovibes.get_cd();
 
-    if (opts.ini_path)
+    if (opts.compute_settings_path)
     {
         try
         {
-            holovibes::api::load_compute_settings(opts.ini_path.value());
+            holovibes::api::load_compute_settings(opts.compute_settings_path.value());
         }
         catch (std::exception&)
         {